# src/main/resources/api/api.yaml
openapi: 3.0.0
info:
  title: Tramways API
  description: Tramways API
  version: 1.0.0
servers:
  - url: http://localhost:8762/tramways/rest

paths:
  # # # # # # #
  # PROJECTS
  # # # # # # #
  /projects:
    post:
      operationId: 'createProject'
      summary: 'Creates a new project'
      tags: ['projects']
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '200':
          description: Ok
    get:
      operationId: 'getProjects'
      summary: 'Gets user projects'
      tags: ['projects']
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDescription'

  /projects/{id}:
    get:
      summary: 'Gets project'
      operationId: 'getProject'
      tags: ['projects']
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    put:
      summary: 'Updates a project'
      operationId: 'updateProject'
      tags: ['projects']
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Ok
    delete:
      summary: 'Deletes a project'
      operationId: 'deleteProject'
      tags: ['projects']
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /projects/{projectId}/maps:
    post:
      summary: 'Creates a map'
      operationId: 'createMap'
      tags: ['projects']
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMapRequest'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoadMap'
  /projects/{projectId}/maps/{mapId}:
    get:
      summary: 'Gets a map'
      operationId: 'getMap'
      tags: ['projects']
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: mapId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoadMap'
    put:
      summary: 'Updates a map'
      operationId: 'updateMap'
      tags: ['projects']
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: mapId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMapRequest'
      responses:
        '200':
          description: Ok
    delete:
      summary: 'Deletes a map'
      operationId: 'deleteMap'
      tags: ['projects']
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: mapId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ok
  /projects/{projectId}/maps/{mapId}/{analysisId}:
    get:
      summary: 'Gets an analysis'
      operationId: 'getAnalysis'
      tags: ['projects']
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: mapId
          required: true
          schema:
            type: string
        - in: path
          name: analysisId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Analysis'
    delete:
      summary: 'Deletes an analysis'
      operationId: 'deleteAnalysis'
      tags: ['projects']
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: mapId
          required: true
          schema:
            type: string
        - in: path
          name: analysisId
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /analysis/available:
    get:
      summary: 'Gets available analysis for map'
      operationId: 'getAvailableAnalysis'
      tags: ['analysis']
      parameters:
        - in: query
          name: projectId
          required: true
          schema:
            type: string
        - in: query
          name: mapId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnalysisType'
  /analysis/launch:
    post:
      summary: 'Launch selected analysis'
      operationId: 'launchAnalysis'
      tags: ['analysis']
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisRequest'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
  /configurations/categories:
    get:
      operationId: 'getConfigurationCategories'
      summary: 'Gets all configuration groups'
      tags: ['configurations']
      responses:
        '200':
          description: 'List of configuration groups'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConfigurableCategory'
  /configurations:
    get:
      operationId: 'getConfigurations'
      summary: 'Get configurations given category'
      tags: ['configurations']
      parameters:
        - in: query
          name: category
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'List of configurations'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemConfiguration'
  /configurations/suggestions:
    get:
      operationId: 'getPropertiesSuggestions'
      summary: 'Get properties given category'
      tags: ['configurations']
      parameters:
        - in: query
          name: category
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'List of properties'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Property'
  /configurations/{category}:
    post:
      operationId: 'addConfiguration'
      summary: 'Adds configuration to a category'
      tags: ['configurations']
      parameters:
        - in: path
          name: category
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConfigurationRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /configurations/{configurationId}:
    put:
      operationId: 'editConfiguration'
      summary: 'Edits a configuration'
      tags: ['configurations']
      parameters:
        - in: path
          name: configurationId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConfigurationRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    delete:
      operationId: 'removeConfiguration'
      summary: 'Removes a configuration'
      tags: ['configurations']
      parameters:
        - in: path
          name: configurationId
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Ok'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateConfigurationRequest:
      type: object
      properties:
        name:
          type: string
        props:
          type: array
          items:
            $ref: '#/components/schemas/Property'
    UpdateConfigurationRequest:
      type: object
      properties:
        name:
          type: string
        props:
          type: array
          items:
            $ref: '#/components/schemas/Property'
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
    CreateMapRequest:
      type: object
      properties:
        map:
          $ref: '#/components/schemas/RoadMap'
    UpdateMapRequest:
      type: object
      properties:
        map:
          $ref: '#/components/schemas/RoadMap'
    UserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRole'
      required: [roles]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    ChangePasswordRequest:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
    User:
      allOf:
        - $ref: '#/components/schemas/Resource'
        - type: object
          properties:
            username:
              type: string
            roles:
              type: array
              items:
                $ref: '#/components/schemas/UserRole'
          required: [username, roles]
    UserRole:
      type: string
      enum: [CLIENT, EXPERT, ADMIN]
    ConfigurableCategory:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    ProjectDescription:
      allOf:
        - $ref: '#/components/schemas/Resource'
        - type: object
          properties:
            name:
              type: string
            roadMaps:
              type: array
              items:
                $ref: '#/components/schemas/RoadMapDescription'
    Project:
      allOf:
        - $ref: '#/components/schemas/Resource'
        - type: object
          properties:
            name:
              type: string
            roadMaps:
              type: array
              items:
                $ref: '#/components/schemas/RoadMapDescription'
    RoadMapDescription:
      allOf:
        - $ref: '#/components/schemas/Resource'
        - type: object
          properties:
            name:
              type: string
    RoadMap:
      allOf:
        - $ref: '#/components/schemas/Resource'
        - type: object
          properties:
            name:
              type: string
            content:
              $ref: '#/components/schemas/RoadMapContent'
            analysis:
              type: array
              items:
                $ref: '#/components/schemas/AnalysisDescription'
    RoadMapContent:
      type: object
      properties:
        points:
          type: array
          items:
            $ref: '#/components/schemas/RelevantPoint'
        lanes:
          type: array
          items:
            $ref: '#/components/schemas/Lane'
    RelevantPoint:
      allOf:
        - $ref: '#/components/schemas/Configurable'
        - type: object
          properties:
            links:
              type: array
              items:
                $ref: '#/components/schemas/CrossingLink'
    Lane:
      allOf:
        - $ref: '#/components/schemas/Configurable'
        - type: object
          properties:
            sourceId:
              type: string
            destinationId:
              type: string
    CrossingLink:
      allOf:
        - $ref: '#/components/schemas/Configurable'
        - type: object
          properties:
            sourceId:
              type: string
            destinationId:
              type: string
    Configurable:
      type: object
      discriminator:
        propertyName: 'configurableType'
      properties:
        id:
          type: string
        category:
          type: string
        configurableType:
          type: string
        props:
          type: array
          items:
            $ref: '#/components/schemas/Property'
    ItemConfiguration:
      allOf:
        - $ref: '#/components/schemas/Resource'
        - type: object
          properties:
            category:
              type: string
            name:
              type: string
            props:
              type: array
              items:
                $ref: '#/components/schemas/Property'
    Property:
      type: object
      discriminator:
        propertyName: propertyType
      properties:
        name:
          type: string
        description:
          type: string
        propertyType:
          type: string
        valid:
          type: boolean
      required: [propertyType]
    IntegerProperty:
      allOf:
        - $ref: '#/components/schemas/Property'
        - type: object
          properties:
            value:
              type: integer
    StringProperty:
      allOf:
        - $ref: '#/components/schemas/Property'
        - type: object
          properties:
            value:
              type: string
    ChoiceProperty:
      allOf:
        - $ref: '#/components/schemas/Property'
        - type: object
          properties:
            choices:
              type: array
              items:
                $ref: '#/components/schemas/ChoiceElement'
            value:
              type: string
    ChoiceElement:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
    DecimalProperty:
      allOf:
        - $ref: '#/components/schemas/Property'
        - type: object
          properties:
            value:
              type: number
    DistributionProperty:
      allOf:
        - $ref: '#/components/schemas/Property'
        - type: object
          properties:
            value:
              $ref: '#/components/schemas/Distribution'
    Distribution:
      type: object
      discriminator:
        propertyName: distributionType
        mapping:
          UNIFORM: '#/components/schemas/UniformDistribution'
          EXPONENTIAL: '#/components/schemas/ExponentialDistribution'
      properties:
        distributionType:
          $ref: '#/components/schemas/DistributionType'
      required: [distributionType]
    DistributionType:
      type: string
      enum: [UNIFORM, EXPONENTIAL]
    UniformDistribution:
      allOf:
        - $ref: '#/components/schemas/Distribution'
        - type: object
          properties:
            left:
              type: number
            right:
              type: number
    ExponentialDistribution:
      allOf:
        - $ref: '#/components/schemas/Distribution'
        - type: object
          properties:
            lambda:
              type: number
    AnalysisType:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    Analysis:
      allOf:
        - $ref: '#/components/schemas/Resource'
        - type: object
          properties:
            name:
              type: string
            status:
              $ref: '#/components/schemas/AnalysisStatus'
            result:
              $ref: '#/components/schemas/AnalysisResult'
    AnalysisDescription:
      allOf:
        - $ref: '#/components/schemas/Resource'
        - type: object
          properties:
            name:
              type: string
            status:
              $ref: '#/components/schemas/AnalysisStatus'
    AnalysisStatus:
      type: string
      enum: [IN_PROGRESS, DONE, ERROR]
    AnalysisResult:
      type: object
      discriminator:
        propertyName: resultType
        mapping:
          XY: '#/components/schemas/XYAnalysisResult'
          STRING: '#/components/schemas/StringAnalysisResult'
      properties:
        resultType:
          $ref: '#/components/schemas/AnalysisResultType'
      required: [distributionType]
    AnalysisResultType:
      type: string
      enum: [XY, STRING]
    XYAnalysisResult:
      allOf:
        - $ref: '#/components/schemas/AnalysisResult'
        - type: object
          properties:
            xLabel:
              type: string
            yLabel:
              type: string
            points:
              type: array
              items:
                $ref: '#/components/schemas/XYPoint'
    StringAnalysisResult:
      allOf:
        - $ref: '#/components/schemas/AnalysisResult'
        - type: object
          properties:
            message:
              type: string
    XYPoint:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
    Resource:
      type: object
      discriminator:
        propertyName: resourceType
      properties:
        uuid:
          type: string
        resourceType:
          type: string
    BooleanWrapper:
      type: object
      properties:
        value:
          type: boolean
    StringWrapper:
      type: object
      properties:
        value:
          type: string
    ErrorWrapper:
      type: object
      properties:
        message:
          type: string
    AnalysisRequest:
      type: object
      properties:
        projectId:
          type: string
        mapId:
          type: string
        analysisTypeId:
          type: string
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/Property'
    AnalysisResponse:
      type: object
      properties:
        warnings:
          type: array
          items:
            type: string
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/Property'
        ok:
          type: boolean
  responses:
    Created:
      description: 'Resource created'
      headers:
        Location:
          schema:
            type: string
    Ok:
      description: 'Request successful'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StringWrapper'
    Unauthorized:
      description: 'You are not authorized'


security:
  - bearerAuth: []